# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AdvicePreviewRequest": {
        "additionalProperties": false,
        "description": "Complete certified input set for one advice preview.\n\nField-for-field the engine's own ``AdviceInputs`` (subclass — nothing is\nredefined), with the gate 6 mapping typed for the wire. A field left\nabsent stays honestly absent and blocks its gate with ``UNEVALUABLE``.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "calculations": {
            "$ref": "#/components/schemas/CalculationStatusesInput"
          },
          "constraints": {
            "$ref": "#/components/schemas/ConstraintsInput"
          },
          "contradictions": {
            "$ref": "#/components/schemas/ContradictionsInput"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "entitlements": {
            "$ref": "#/components/schemas/EntitlementsInput"
          },
          "evidence_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "explanation_facts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Explanation Facts",
            "type": "array"
          },
          "horizon": {
            "minLength": 1,
            "title": "Horizon",
            "type": "string"
          },
          "input_snapshot_id": {
            "minLength": 1,
            "title": "Input Snapshot Id",
            "type": "string"
          },
          "instrument": {
            "$ref": "#/components/schemas/InstrumentResolutionInput"
          },
          "instrument_id": {
            "minLength": 1,
            "title": "Instrument Id",
            "type": "string"
          },
          "limitations": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "liquidity": {
            "$ref": "#/components/schemas/LiquidityInput"
          },
          "portfolio_risk": {
            "$ref": "#/components/schemas/PortfolioRiskInput"
          },
          "probability": {
            "$ref": "#/components/schemas/ProbabilityInput"
          },
          "probability_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Probability Evidence"
          },
          "risk_summary": {
            "minLength": 1,
            "title": "Risk Summary",
            "type": "string"
          },
          "scenario_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Scenario Ids",
            "type": "array"
          },
          "session_event": {
            "$ref": "#/components/schemas/SessionEventInput"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotInput"
          },
          "supersedes": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Supersedes"
          },
          "valid_until": {
            "format": "date-time",
            "title": "Valid Until",
            "type": "string"
          }
        },
        "required": [
          "instrument_id",
          "as_of",
          "valid_until",
          "input_snapshot_id",
          "horizon",
          "direction",
          "risk_summary"
        ],
        "title": "AdvicePreviewRequest",
        "type": "object"
      },
...
      "CalculationStatusesInput": {
        "additionalProperties": false,
        "description": "Wire form of the gate 6 facts.\n\nNarrows the engine's ``Any``-valued mapping to ``CalculationStatus``\nmembers so JSON input (``{\"iv_surface\": \"OK\"}``) reaches the gate as\ncanonical enum values. An absent mapping stays ``None`` (fail-closed at\nthe gate), never an empty default.",
        "properties": {
          "calculation_statuses": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/CalculationStatus"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Calculation Statuses"
          }
        },
        "title": "CalculationStatusesInput",
        "type": "object"
      },
...
      "EngineInfoResponse": {
        "additionalProperties": false,
        "description": "Engine and contract versions backing every verdict. Carries no secret.\n\n``contracts_version`` equals the ``ENGINE_VERSION`` stamp because the\ncanonical contracts are versioned by the same identifier that is recorded\nin every calculation and advice contract (``vertex_core.version``).",
        "properties": {
          "contracts_version": {
            "minLength": 1,
            "title": "Contracts Version",
            "type": "string"
          },
          "engine_version": {
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "gate_versions": {
            "additionalProperties": true,
            "title": "Gate Versions",
            "type": "object"
          }
        },
        "required": [
          "engine_version",
          "contracts_version",
          "gate_versions"
        ],
        "title": "EngineInfoResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulationPreviewRequest": {
        "additionalProperties": false,
        "description": "One declared structure plus its assumptions.\n\nLegs ARE the engine's own :class:`OptionLeg` contract (strict: signed\nnon-zero quantity, strike required on CALL/PUT and forbidden on STOCK,\nnon-negative decimal premium, positive integer multiplier) — the wire\nnever redefines the calculation contract.",
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/SimulationAssumptions"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/OptionLeg"
            },
            "maxItems": 8,
            "minItems": 1,
            "title": "Legs",
            "type": "array"
          }
        },
        "required": [
          "legs",
          "assumptions"
        ],
        "title": "SimulationPreviewRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local, analysis-only API of Vertex One. It serves the canonical contracts of vertex_core and delegates every verdict to the single AdviceEngine. No transactional capability exists; the human remains the sole decision maker, acting outside Vertex.",
    "title": "Vertex One API",
    "version": "vertex_core@0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/advice/preview": {
      "post": {
        "description": "Return the canonical ``AdviceResult`` for certified gate inputs.\n\nPass-through only: the route validates the wire payload and hands it to\nthe single ``AdviceEngine``. Decimals serialize as strings and datetimes\nas ISO-8601 UTC. The result is analytical — the human decides outside\nVertex.",
        "operationId": "post_advice_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvicePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdviceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Evaluate gate inputs through the single AdviceEngine"
      }
    },
    "/api/v1/analysis/{instrument}": {
      "get": {
        "description": "Serve the LAST ``analysis/{instrument}`` snapshot exactly as persisted.\n\nThe API relays the worker's published dossier — the canonical\n``AdviceResult`` of the single ``AdviceEngine`` with its ten gates and\nreason codes, the validated OHLCV bars, the fusion evidence rail and the\n``THEORETICAL`` scenario block (or its typed absence reason) — and\ncomputes nothing. With no snapshot ever published for this instrument\nthe answer is a 200 with ``state = \"empty\"``: absent stays absent,\nnothing is invented.",
        "operationId": "get_analysis",
        "parameters": [
          {
            "in": "path",
            "name": "instrument",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Instrument",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last published analysis dossier for one instrument (or honest empty state)"
      }
    },
...
    "/api/v1/events/stream": {
      "get": {
        "description": "Stream head-version change signals (database polling, coalesced).",
        "operationId": "get_events_stream",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "text/event-stream of `snapshot` events (`{\"resource\": \"<kind>/<key>\", \"version\": <int>}`) and keepalive `ping` events. Signal only — no business data; clients refetch through the REST endpoints."
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Signal-only SSE: snapshot head version changes and pings"
      }
    },
...
    "/api/v1/markets/overview": {
      "get": {
        "description": "Serve the LAST ``markets_overview/global`` snapshot exactly as persisted.\n\nThe API relays the worker's published content — population (``SYNTHETIC``\nshown as-is), sectors/tickers with their server-computed returns and\nweights, breadth, coverage account and the deterministic conclusion — and\ncomputes nothing. With no snapshot ever published the answer is a 200\nwith ``state = \"empty\"``: absent stays absent, nothing is invented.",
        "operationId": "get_markets_overview",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketsOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Last published markets overview snapshot (or honest empty state)"
      }
    },
    "/api/v1/options/{underlying}/chain": {
      "get": {
        "description": "Serve the LAST ``option_chain/{underlying}`` snapshot exactly as persisted.\n\nThe API relays the worker's published content — per-(expiration,\ntrading_class) groups with complete contract identities, verbatim quotes\nand their quality, the worker's Vertex IV/Greeks (``THEORETICAL``, with\ntheir ``CalculationRecord`` lineage) or their typed refusal reasons, the\ncoverage account and the displayed row budget — and computes nothing.\nWith no snapshot ever published for this underlying the answer is a 200\nwith ``state = \"empty\"``: absent stays absent, nothing is invented.",
        "operationId": "get_option_chain",
        "parameters": [
          {
            "in": "path",
            "name": "underlying",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Underlying",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Last published option chain for one underlying (or honest empty state)"
      }
    },
    "/api/v1/simulations/preview": {
      "post": {
        "description": "Run one THEORETICAL simulation preview through vertex_core.\n\nOrchestration only: the mandatory ``defined_risk_check``, the exact\n``payoff_at_expiry``, the authority-certified breakevens and the bounded\n``scenario_grid`` all run inside ``vertex_core.calculations.options`` on\na worker thread (``run_in_threadpool`` — the event loop never computes).\nNothing is persisted; nothing here is, or ever becomes, an order.",
        "operationId": "post_simulations_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          },
          "422": {
            "description": "Rejected fail-closed with the exact machine-readable reason: either a wire-contract violation, or the defined-risk verifier's code (e.g. OUTSIDE_CLOSED_CATALOG, UNCOVERED_SHORT_UPSIDE_TAIL, VERTICAL_DEBIT_NOT_BELOW_WIDTH), or a typed calculation-domain violation from vertex_core."
          }
        },
        "summary": "THEORETICAL preview of one declared structure (no persistence)"
      }
    },
    "/api/v1/system/capabilities": {
      "get": {
        "description": "Cross the FULL declared manifest with the latest persisted probes.\n\nEvery manifest entry is present (``total`` equals the manifest size); a\ncapability never probed answers ``ERROR`` with reason ``NEVER_TESTED``.\nHealth blocks report the database (``SELECT 1``), both snapshot heads,\nand the worker through the explicitly labeled ``heartbeat_proxy``.",
        "operationId": "get_system_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemCapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Declared capabilities crossed with really-probed statuses, plus health"
      }
    },
    "/api/v1/system/engine": {
      "get": {
        "description": "Report the engine, contract and per-gate versions. Never a secret.",
        "operationId": "get_system_engine",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Engine and contract versions"
      }
    },
    "/api/v1/today/attention": {
      "get": {
        "description": "Serve the LAST ``attention/global`` snapshot exactly as persisted.\n\nThe API relays the worker's published content — population (``SYNTHETIC``\nshown as-is), coverage, items with provenance — and computes nothing.\nWith no snapshot ever published the answer is a 200 with ``state =\n\"empty\"``: absent stays absent, nothing is invented.",
        "operationId": "get_today_attention",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required: no valid WebAuthn session cookie (or missing/invalid CSRF header on a mutation). Always the same generic body with detail code AUTH_REQUIRED."
          }
        },
        "summary": "Last published attention snapshot (or honest empty state)"
      }
    }
  }
}